==> Code/1_Explore_Data.Rmd <==
---
title: "Calculating life expectancy and exploring missingness"
author: "Corinne Riddell"
date: "July 22, 2016"
output: html_document
---

```{r load_libraries, echo=F, warning=F, message=F}
library(tidyverse)
```

*About the dataset*
The original dataset was collected by Sam using software from the National Cancer Institute's Surveillance, Epidemiology, and End Results Program (known as the SEER*Stat Software) to access the National Vital Statistics System that is maintained by the National Center for Health Statistics.

```{r read_raw_data}
dat <- read.table("./Data/deaths-cause-6913.txt", header = F, sep="\t")
#see ./Data/deaths-cause-6913.dic for the data dictionary.

names(dat) <- c("state.n", "age.n", "sex.n", "race.n", "COD.n", "year.0", "crude.rate", "deaths", "population") 
```

```{r recode_variables}
dat <- dat %>%
  mutate(state = factor(state.n, levels = c(0:50), labels = c("Alabama", "Alaska", "Arizona", "Arkansas","California",
                                                             "Colorado", "Connecticut", "Delaware", "Washington DC",
                                                             "Florida", "Georgia", "Hawaii","Idaho", "Illinois", 
                                                             "Indiana","Iowa", "Kansas", "Kentucky", "Louisiana", 
                                                             "Maine", "Maryland", "Massachusetts", "Michigan", 
                                                             "Minnesota", "Mississippi","Missouri", "Montana", 
                                                             "Nebraska", "Nevada", "New Hampshire", "New Jersey",
                                                             "New Mexico", "New York","North Carolina","North Dakota",
                                                             "Ohio", "Oklahoma","Oregon", "Pennsylvania", 
                                                             "Rhode Island", "South Carolina", "South Dakota", 
                                                             "Tennessee", "Texas", "Utah", "Vermont", "Virginia",
                                                             "Washington", "West Virginia", "Wisconsin", "Wyoming")), 
         age = factor(age.n, levels = c(0:19), labels = c("<1 year", "1-4 years", "5-9 years", "10-14 years", 
                                                         "15-19 years", "20-24 years", "25-29 years", "30-34 years",
                                                         "35-39 years", "40-44 years", "45-49 years", "50-54 years", 
                                                         "55-59 years", "60-64 years", "65-69 years", "70-74 years", 
                                                         "75-79 years", "80-84 years","85+ years", "Unknown"), ordered = T),
         age_minbin = 0*(age == "<1 year") + 1*(age == "1-4 years") + 5*(age == "5-9 years") +
                10*(age == "10-14 years") + 15*(age == "15-19 years") + 20*(age == "20-24 years") + 
                25*(age == "25-29 years") + 30*(age == "30-34 years") + 35*(age == "35-39 years") +
                40*(age == "40-44 years") + 45*(age == "45-49 years") + 50*(age == "50-54 years") + 
                55*(age == "55-59 years") + 60*(age == "60-64 years") + 65*(age == "65-69 years") +   
                70*(age == "70-74 years") + 75*(age == "75-79 years") + 80*(age == "80-84 years") + 
                85*(age == "85+ years"),
         age_bin_size = 1*(age == "<1 year") + 4*(age == "1-4 years") + 5*(age != "<1 year" & age != "1-4 years"),
         sex = factor(sex.n, levels = c(0, 1), labels = c("Male", "Female")),
         race = factor(race.n, levels = c(0, 1), labels = c("White", "Black")),
         COD = factor(COD.n, levels = c(0:5), labels = c("Cardiovascular", "Cancers", "Communicable", 
                                                        "Non-communicable", "Injuries", "All other causes")), 
         year = factor(year.0, levels = c(0:44), labels = as.character(1969:2013)),
         year.n = as.numeric(as.character(year)),
         race.sex = interaction(race, sex, sep = " "),
         state.year.sex = interaction(state, year, sex),
         state.year.race.sex = interaction(state, year, race, sex))

head(dat)

dat$stabbrs <- state.abb[match(dat$state, state.name)]
unique(dat$state[is.na(dat$stabbrs)==T]) #Washington DC not defined
dat$stabbrs[dat$state == "Washington DC"] <- "DC"

dat$stabbrs <- factor(dat$stabbrs)

dat$COD.abbrev <- "CAR"
dat$COD.abbrev[dat$COD == "Cancers"] <- "CAN"
dat$COD.abbrev[dat$COD == "Communicable"] <- "CDS"
dat$COD.abbrev[dat$COD == "Non-communicable"] <- "NDS"
dat$COD.abbrev[dat$COD == "Injuries"] <- "INJ"
dat$COD.abbrev[dat$COD == "All other causes"] <- "AOC"

dat$RID <- paste0(dat$stabbrs, dat$year, substr(dat$sex, 1, 1),
                 substr(dat$race, 1, 1), dat$age.n, dat$COD.abbrev)
```

```{r remove_unknown_age_group}
#there is an unknown age group for each combo of state-year-race-sex-age:
table(dat$age) 

#Count of death is always equal to 0 for unknown age groups
table(dat$age, dat$deaths==0)

#remove these data from the data we analyze:
dat.clean <- dat[dat$age != "Unknown", ]
```

##Population counts of 0

It is important to explore how many cells have population counts of zero because these break the Bayesian crank, and also currently breaks the life expectancy function.

```{r recode_pops_of_zero}
table(dat.clean$population == 0, useNA = "always")
#654 cells have Populations of 0.

table(dat.clean$deaths[dat.clean$population == 0] == 0, useNA = "always")
#Of these, all have Counts of 0, except for 26 with counts of NA

#replace the NA Counts with 0 for the 26 cells with Populations of 0.
dat.clean$deaths[is.na(dat.clean$deaths) == T & dat.clean$population == 0] <- 0
summary(dat.clean$deaths[dat.clean$population == 0])

summary(dat.clean$crude.rate[dat.clean$population == 0])
#whenever the population is 0, the crude.rate is equal to NA.
#need to think: should these be equal to 0, or NA? How will it affect the calculations?
```

```{r list_states_0_le20_popcounts}
#States with at least one strata of 0
na.pops <- dat.clean %>% 
  select(state, population) %>% 
  filter(population == 0)

unique(na.pops$state)

pops.le.20 <- dat.clean %>% select(state, population) %>% filter(population <= 20)
unique(pops.le.20$state)
```

**Missingness of Count (of deaths)**
If there are between 1 and 9 deaths in the strata, this information is 'coarsened' and the number of deaths (Count) is coded as NA. Later in this document we investigate trends in supression/missingness to understand how it differs according to state, age, race, and sex.

Since our goal is to compute the difference in life expectancy between Blacks and Whites by strata, we impute death counts at their logical extremes to explore how this affects the results. Further, we use this rough data to create the shiny app while the Bayesian model is being developed/run.

```{r fill_in_missing_counts}
dat.clean <- dat.clean %>% mutate(deaths_md1 = ifelse(is.na(deaths)==T, pmin(1, population), deaths),
                                  deaths_md5 = ifelse(is.na(deaths)==T, pmin(5, population), deaths),
                                  deaths_md9 = ifelse(is.na(deaths)==T, pmin(9, population), deaths)
                                  )

#check recoding
dat.clean$deaths_grouped <- cut(dat.clean$deaths, c(-1, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 22040))
dat.clean$deaths1_grouped <- cut(dat.clean$deaths_md1, c(-1, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 22040))
dat.clean$deaths5_grouped <- cut(dat.clean$deaths_md5, c(-1, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 22040))
dat.clean$deaths9_grouped <- cut(dat.clean$deaths_md9, c(-1, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 22040))

table(dat.clean$deaths_grouped, dat.clean$deaths1_grouped, useNA = "always")
table(dat.clean$deaths_grouped, dat.clean$deaths5_grouped, useNA = "always")
table(dat.clean$deaths_grouped, dat.clean$deaths9_grouped, useNA = "always")

dat.clean <- dat.clean[ , !names(dat.clean) %in% c("deaths_grouped", "deaths1_grouped", "deaths5_grouped", "deaths9_grouped")]
```

```{r addvar_cod_proportions} 
# Add the proportions of deaths due to each cause as a new variable.
# This variable is used to decompose the life expectancy gap by cause of death.

dat.clean <- dat.clean %>% group_by(state, race.sex, year.n, age_minbin) %>%
                           arrange(state, race.sex, year.n, age_minbin) %>%
                           mutate(total_deaths1 = sum(deaths_md1),
                                  total_deaths5 = sum(deaths_md5),
                                  total_deaths9 = sum(deaths_md9),
                                  cod_prop_death1 = deaths_md1/total_deaths1, 
                                  cod_prop_death5 = deaths_md5/total_deaths5,
                                  cod_prop_death9 = deaths_md9/total_deaths9)

length(dat.clean$total_deaths5[dat.clean$total_deaths5 == 0])
summary(dat.clean$cod_prop_death5[dat.clean$total_deaths5 == 0])
#76998 strata with 0 deaths from any cause (12833 when divided by 6 -- the number of causes)
#7% of strata have no deaths from any cause.
```

To calculate life expectancy, we have to aggregate the data across causes of death.

```{r aggregate_over_COD}
dat.aggregated <- dat.clean %>% group_by(state, year, race.sex, age) %>%
                                arrange(state, race.sex, year, age) %>%
                                summarise(population = first(population), 
                                          race = first(race), 
                                          sex = first(sex), 
                                          state.year.race.sex = first(state.year.race.sex),
                                          age_bin_size = first(age_bin_size),
                                          age_minbin = first(age_minbin),
                                          year.n = first(year.n),
                                          total_deaths1 = first(total_deaths1),
                                          total_deaths5 = first(total_deaths5),
                                          total_deaths9 = first(total_deaths9),
                                          cod_prop_death1 = first(cod_prop_death1),
                                          cod_prop_death5 = first(cod_prop_death5),
                                          cod_prop_death9 = first(cod_prop_death9))

#return "stratum ids" with any age-specific population counts of 0.
non.zero.stratums <- setdiff(levels(dat.aggregated$state.year.race.sex), 
                             unique(dat.aggregated$state.year.race.sex[dat.aggregated$population == 0]))
```

Add the total population (across age brackets) to the dat.aggregated dataset.

```{r create_var_pop_across_age}
dat3 <- dat.aggregated %>% ungroup() %>% 
                           group_by(state.year.race.sex) %>%
                           summarise(pop_across_age = sum(population))

dat.aggregated <- merge(dat.aggregated, dat3, by = "state.year.race.sex")

rm(dat3)

dat.aggregated <- dat.aggregated %>% arrange(state, year.n, race, sex, age_minbin)
```

```{r write_data_to_csvs}
write_csv(dat.aggregated, path = "./dat_aggregated.csv", col_names = T)

excluded.states <- c("Alaska", "Hawaii", "Idaho", "Maine", "Montana", 
                     "New Hampshire", "North Dakota", "South Dakota", "Utah", "Vermont", "Wyoming")

included.states <- setdiff(unique(levels(dat.clean$state)), excluded.states)

for(i in 1:length(included.states)) { 
  name = paste0('dat_clean_', included.states[i], '.csv')
  ds_temp = dat.clean[dat.clean$state == included.states[i], ]
  write.csv(ds_temp, file = paste0("./Data/", name))
}
rm(ds_temp)
```
