==> Code/3_combine_results.Rmd <==
---
title: "Combine_results"
date: "2/9/2017"
output: html_document
---

**Combine results across the sets**
The analysis was run in 3 sets of states. In this file we combine the results and tidy the data for graphing and presentation.

```{r initialize}
library(stringr)
library(dplyr)
source("./Code/life_expectancy_functions.R")
```

```{r}
included.states <- c("Alabama", "Arizona", "Arkansas", "California", "Colorado", "Connecticut",
"Delaware", "Florida", "Georgia", "Illinois", "Indiana", "Iowa", "Kansas", "Kentucky",
"Louisiana", "Maryland", "Massachusetts", "Michigan", "Minnesota", "Mississippi",   
"Missouri", "Nebraska", "Nevada", "New Jersey", "New Mexico", "New York",
"North Carolina", "Ohio", "Oklahoma", "Oregon", "Pennsylvania", "Rhode Island",
"South Carolina",  "Tennessee", "Texas", "Virginia", "Washington", "Washington DC",
"West Virginia", "Wisconsin") 
```

```{r combine_results_seed123}
age_cod_results <- read.csv(paste0("./Results2/age_cod_", included.states[1], "_123.csv"))
age_decomp_results <- read.csv(paste0("./Results2/age_decomp_", included.states[1], "_123.csv"))
BlackWhite_results <- read.csv(paste0("./Results2/BlackWhite_", included.states[1], "_123.csv"))
cod_decomp_results <- read.csv(paste0("./Results2/cod_marginal_", included.states[1], "_123.csv"))
cod_change_results <- read.csv(paste0("./Results2/cod_change_", included.states[1], "_123.csv"))

mortality.rates <- read.csv(paste0("./Results2/mortality_rates_", included.states[1], "_123.csv"))
mortality.rates.diff <- read.csv(paste0("./Results2/mortality_rates_diff_", included.states[1], "_123.csv"))

for(state_i in 2:length(included.states)){
  
    age_cod_results <- rbind(age_cod_results, read.csv(paste0("./Results2/age_cod_", included.states[state_i], "_123.csv")))
  
    age_decomp_results <- rbind(age_decomp_results, read.csv(paste0("./Results2/age_decomp_", included.states[state_i], "_123.csv")))

    BlackWhite_results <- rbind(BlackWhite_results, read.csv(paste0("./Results2/BlackWhite_", included.states[state_i], "_123.csv")))
  
    cod_decomp_results <- rbind(cod_decomp_results, read.csv(paste0("./Results2/cod_marginal_", included.states[state_i], "_123.csv")))    
    cod_change_results <- rbind(cod_change_results, read.csv(paste0("./Results2/cod_change_", included.states[state_i], "_123.csv")))    
    mortality.rates <- rbind(mortality.rates, read.csv(paste0("./Results2/mortality_rates_", included.states[state_i], "_123.csv")))
  
    mortality.rates.diff <- rbind(mortality.rates.diff, read.csv(paste0("./Results2/mortality_rates_diff_", included.states[state_i], "_123.csv")))
}

mortality.rates <- mortality.rates %>% mutate(state.reorder = reorder(state, as.numeric(Census_Division)))
mortality.rates.diff <- mortality.rates.diff %>% mutate(state.reorder = reorder(state, as.numeric(Census_Division)))
```

```{r tidy_data, echo=FALSE}
ncols <- str_split_fixed(as.character(BlackWhite_results$stratum.id), "\\.", 3)
BlackWhite_results$state <- as.factor(ncols[ , 1])
BlackWhite_results$year <- as.numeric(ncols[ , 2])
BlackWhite_results$sex <- as.factor(ncols[ , 3])

age_decomp_results <- merge(BlackWhite_results[, c("stratum.id", "state", "year", "sex")], age_decomp_results, by = "stratum.id")
cod_decomp_results <- merge(BlackWhite_results[, c("stratum.id", "state", "year", "sex")], cod_decomp_results, by = "stratum.id")
age_cod_results <- merge(BlackWhite_results[, c("stratum.id", "state", "year", "sex")], age_cod_results, by = "stratum.id")

age_cod_results$COD <- factor(age_cod_results$COD, levels = 1:6, 
                               labels = c("All other causes", "Cancers", "Cardiovascular", "Communicable",
                                          "Injuries", "Non-communicable"))

age_cod_results$age <- factor(age_cod_results$age_minbin, 
                              levels = c(0, 1, seq(5, 85, by = 5)),
                              labels = c("<1 year", "1-4 years", "5-9 years", "10-14 years", 
                                         "15-19 years", "20-24 years", "25-29 years", "30-34 years",
                                         "35-39 years", "40-44 years", "45-49 years", "50-54 years", 
                                         "55-59 years", "60-64 years", "65-69 years", "70-74 years", 
                                         "75-79 years", "80-84 years","85+ years"), ordered = T)

age_decomp_results$age <- factor(age_decomp_results$age_minbin, 
                                 levels = c(0, 1, seq(5, 85, by = 5)),
                                 labels = c("<1 year", "1-4 years", "5-9 years", "10-14 years", 
                                            "15-19 years", "20-24 years", "25-29 years", "30-34 years",
                                            "35-39 years", "40-44 years", "45-49 years", "50-54 years", 
                                            "55-59 years", "60-64 years", "65-69 years", "70-74 years", 
                                            "75-79 years", "80-84 years","85+ years"), ordered = T)
```

```{r add_Census_Variables, echo = F}
BlackWhite_results <- add_Census_Region(BlackWhite_results, state)
BlackWhite_results <- add_Census_Division(BlackWhite_results, state)
BlackWhite_results <- reorder.as.map(BlackWhite_results, "state", "stabbrs")

BlackWhite_results$stabbrs <- state.abb[match(BlackWhite_results$state, state.name)]
BlackWhite_results$stabbrs[BlackWhite_results$state == "Washington DC"] <- "DC"

cod_decomp_results <- merge(BlackWhite_results %>% select(stratum.id, Census_Region, Census_Division, stabbrs), 
                            cod_decomp_results, by = "stratum.id", all.y = T)
cod_decomp_results <- reorder.as.map(cod_decomp_results, "state", "stabbrs")

```

```{r prep_decomp_for_plotting, echo = F}
cod_decomp_results$widened_gap <- sign(cod_decomp_results$COD_cont_yrs_mean)
cod_decomp_results <- make_pretty_decomp_plot(decomp.table = cod_decomp_results, strat.var.1 = "stratum.id", 
                                   partition.bar.var = "COD", sign.var = "widened_gap",
                                   decomp.var = "COD_cont_yrs_mean", decomp.var.prop = "COD_cont_prop_mean",
                                   type.of.data = "practice")

age_decomp_results$widened_gap <- sign(age_decomp_results$age_cont_yrs_mean)
age_decomp_results <- make_pretty_decomp_plot(decomp.table = age_decomp_results, strat.var.1 = "stratum.id", 
                                   partition.bar.var = "age_minbin", sign.var = "widened_gap",
                                   decomp.var = "age_cont_yrs_mean", decomp.var.prop = "age_cont_prop_mean",
                                   type.of.data = "practice")

age_cod_results$widened_gap <- sign(age_cod_results$age_COD_cont_yrs_mean)
age_cod_results <- make_pretty_decomp_plot(decomp.table = age_cod_results, strat.var.1 = "stratum.id",
                                           strat.var.2 = "state", strat.var.3 = "age_minbin",
                                           partition.bar.var = "COD", sign.var = "widened_gap",
                                           decomp.var = "age_COD_cont_yrs_mean", decomp.var.prop = "age_COD_prop_mean",
                                           type.of.data = "practice")
```

```{r split-age-cod-data-for-save}
age_cod_results_female <- age_cod_results %>% filter(sex == "Female")
age_cod_results_male <- age_cod_results %>% filter(sex == "Male")
```

```{r save_combined_datasets, echo = F}
write.csv(age_cod_results_female, "./Results2/age_cod_results_female.csv")
write.csv(age_cod_results_male, "./Results2/age_cod_results_male.csv")
write.csv(age_decomp_results, "./Results2/age_decomp_results.csv")
write.csv(cod_decomp_results, "./Results2/cod_decomp_results.csv")
write.csv(BlackWhite_results, "./Results2/BlackWhite_results.csv")
write.csv(cod_change_results, "./Results2/cod_change_results.csv")
write.csv(mortality.rates, "./Results2/mortality_rates_combined.csv")
write.csv(mortality.rates.diff, "./Results2/mortality_rates_diff_combined.csv")
```
